requirements.txt
Django==2.2.6
djangorestframework==3.10.3
drf-generators==0.4.0
mysqlclient==1.4.4
orator==0.9.9
